refactor(api): split AppDependencies into per-domain interfaces
Moves the 138-line AppDependencies bag from src/server/index.ts into
src/server/deps.ts as named per-domain slices:

  DbDeps, SettingsDeps, UserDeps, PipelineDeps, JobDeps,
  RecommendationDeps, SubscriptionDeps, TargetDeps, LibraryDeps,
  SlskdDeps, DashboardDeps, DiscoveryDeps, OptionalRouteDeps.

AppDependencies is re-exported as the intersection of every slice, so
existing route factories that accept the full bag keep working without
changes. Future route-factory refactors can now narrow to just the
slice they consume - a lighter contract for their tests.

Addresses audit task 9.12.
