+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "digarr",
-  "version": "0.36.4",
+  "version": "0.36.5",
   "type": "module",
   "private": true,
   "scripts": {
