+++ src/server/index.ts
@@ -5,37 +5,6 @@ import { cors } from 'hono/cors'
 import { HTTPException } from 'hono/http-exception'
 import { secureHeaders } from 'hono/secure-headers'
 import { envConfig } from '@/config/env'
-import type { OidcService } from '@/core/auth/oidc'
-import type { DiscoveryModeRegistry } from '@/core/discovery-modes/registry'
-import type { DiscoveryModeRequest } from '@/core/discovery-modes/request'
-import type { GenreService } from '@/core/genre/service'
-import type { SupportedLocale } from '@/core/i18n/locales'
-import type { AlbumCoverage } from '@/core/library/album-coverage'
-import type { LibraryHealthService } from '@/core/library/health'
-import type { SkyHookWarmer } from '@/core/library/skyhook-warmer'
-import type { LibrarySyncStore } from '@/core/library/store'
-import type { SyncOrchestrator } from '@/core/library/sync'
-import type { PipelineOrchestrator } from '@/core/pipeline/orchestrator'
-import type { SubscriptionScheduler } from '@/core/pipeline/subscription-scheduler'
-import type { AiProviderRegistry } from '@/core/providers/registry'
-import type { ServiceTestResult } from '@/core/types'
-import type { ArtistRow } from '@/db/queries/artists'
-import type { BatchRow } from '@/db/queries/batches'
-import type { ActivityEntry, TasteGenre } from '@/db/queries/dashboard'
-import type {
-  ListRecommendationsFilters,
-  ListRecommendationsResult,
-  RecommendationWithArtist,
-  StatusUpdateExtra,
-} from '@/db/queries/recommendations'
-import type { SettingsRow, SetupConfig } from '@/db/queries/settings'
-import type { SubscriptionInsert, SubscriptionUpdate } from '@/db/queries/subscriptions'
-import type { subscriptions } from '@/db/schema'
-
-type SubscriptionRow = typeof subscriptions.$inferSelect
-
-import type { TargetInsert, TargetRow, TargetUpdate } from '@/db/queries/targets'
-import type { UserPublic } from '@/db/queries/users'
 import { VERSION } from '@/version'
 import { problem } from './helpers/problem'
 import { adminGuard } from './middleware/admin-guard'
@@ -62,10 +31,8 @@ import { moodRoutes } from './routes/mood'
 import { oauthRoutes } from './routes/oauth'
 import { oidcRoutes } from './routes/oidc'
 import { pipelineRoutes } from './routes/pipeline'
-import type { PlaylistDeps } from './routes/playlists'
 import { playlistRoutes } from './routes/playlists'
 import { recommendationRoutes } from './routes/recommendations'
-import type { SearchDeps } from './routes/search'
 import { searchRoutes } from './routes/search'
 import { settingsRoutes } from './routes/settings'
 import { setupRoutes } from './routes/setup'
@@ -73,147 +40,16 @@ import { slskdRoutes } from './routes/slskd'
 import { subscriptionRoutes } from './routes/subscriptions'
 import { targetRoutes } from './routes/targets'
 import { userRoutes } from './routes/users'
-import type { DiscoveryConnectionSnapshot, HonoEnv } from './types'
+import type { HonoEnv } from './types'
 
-export type AppDependencies = {
-  db: import('@/db').Database
-  storeDb: import('@/core/pipeline/store').StoreDb
-  orchestrator: PipelineOrchestrator
-  scheduler: SubscriptionScheduler
-  providerRegistry: AiProviderRegistry
-  isSetupComplete: () => Promise<boolean>
-  getSettings: () => Promise<SettingsRow | null>
-  updateSettings: (partial: Record<string, unknown>) => Promise<void>
-  completeSetup: (config: SetupConfig) => Promise<unknown>
-  // Pipeline status
-  getLastBatch: () => Promise<{ id: number; createdAt: Date | string; status: string } | null>
-  // Recommendation query functions
-  listRecommendations: (filters?: ListRecommendationsFilters) => Promise<ListRecommendationsResult>
-  getRecommendation: (id: number) => Promise<RecommendationWithArtist | null>
-  updateRecommendationStatus: (
-    id: number,
-    status: string,
-    extra?: StatusUpdateExtra,
-  ) => Promise<void>
-  bulkUpdateStatus: (ids: number[], status: string) => Promise<void>
-  filterOwnedIds: (ids: number[], userId: number | undefined) => Promise<number[]>
-  // Batch query functions
-  listBatches: () => Promise<BatchRow[]>
-  getBatch: (id: number) => Promise<BatchRow | null>
-  // Artist query functions
-  getArtistById: (id: number) => Promise<ArtistRow | null>
-  restartScheduler: (cron: string | null) => void
-  restartPlaylistScheduler: () => Promise<void>
-  restartLibraryMaintenanceScheduler?: (intervalHours: number) => void
-  // User query functions
-  createUser: (data: {
-    username: string
-    passwordHash: string
-    isAdmin?: boolean
-  }) => Promise<UserPublic>
-  getUserByUsername: (
-    username: string,
-  ) => Promise<{ id: number; username: string; passwordHash: string; isAdmin: boolean } | null>
-  getUserById: (id: number) => Promise<UserPublic | null>
-  getUserCount: () => Promise<number>
-  updatePassword: (id: number, passwordHash: string) => Promise<void>
-  updateUserPreferredLocale: (id: number, preferredLocale: SupportedLocale | null) => Promise<void>
-  // OIDC + user management
-  getOidcService: () => Promise<OidcService | null>
-  getUserByOidcSubject: (subject: string) => Promise<{ id: number; username: string } | null>
-  getUserByEmail: (email: string) => Promise<{ id: number; username: string } | null>
-  updateUser: (
-    id: number,
-    data: { isAdmin?: boolean; email?: string; oidcSubject?: string },
-  ) => Promise<void>
-  listUsers: () => Promise<UserPublic[]>
-  deleteUser: (id: number) => Promise<void>
-  // Genre service
-  genreService: GenreService
-  // Library health service
-  libraryHealth: LibraryHealthService
-  // SkyHook cache warmer (optional - absent if Lidarr is not configured)
-  skyhookWarmer?: SkyHookWarmer | null
-  // Library sync orchestrator + store
-  librarySync: SyncOrchestrator
-  librarySyncStore: LibrarySyncStore
-  slskdOrchestrator?: {
-    readonly isSyncing: boolean
-    triggerSync: () => Promise<void>
-    warmup: () => Promise<void>
-    getActiveJobs: (limit?: number) => Promise<
-      Array<{
-        id: number
-        targetId: number
-        recommendationId: number | null
-        state: string
-        releaseTitle: string
-      }>
-    >
-  }
-  albumCoverage?: {
-    getCoverageForArtist: (userId: number, artistMbid: string) => Promise<AlbumCoverage>
-  }
-  // Subscription query functions
-  subscriptionQueries: {
-    createSubscription: (data: SubscriptionInsert) => Promise<SubscriptionRow>
-    getSubscription: (id: number) => Promise<SubscriptionRow | null>
-    getSubscriptionsByUser: (userId: number) => Promise<SubscriptionRow[]>
-    getEnabledSubscriptions: () => Promise<SubscriptionRow[]>
-    updateSubscription: (id: number, data: SubscriptionUpdate) => Promise<void>
-    deleteSubscription: (id: number) => Promise<void>
-  }
-  // Manual subscription trigger
-  runSubscription: (id: number) => Promise<void>
-  // Target management
-  targetQueries: {
-    createTarget: (data: TargetInsert) => Promise<{ id: number }>
-    getTargetsByUser: (userId: number) => Promise<TargetRow[]>
-    getAllTargets: () => Promise<TargetRow[]>
-    getTarget: (id: number) => Promise<TargetRow | null>
-    updateTarget: (id: number, data: TargetUpdate) => Promise<void>
-    deleteTarget: (id: number) => Promise<void>
-  }
-  testTargetConnection: (
-    type: string,
-    config: Record<string, unknown>,
-  ) => Promise<ServiceTestResult>
-  getEnabledTargetsForUser: (
-    userId: number,
-  ) => Promise<import('@/core/targets/types').DestinationTarget[]>
-  getFeedbackHistory: () => Promise<Map<string, { approved: number; total: number }>>
-  dashboardQueries: {
-    getTopGenresForUser: (userId: number | undefined) => Promise<TasteGenre[]>
-    getRecentActivity: (
-      userId: number | undefined,
-      isAdmin: boolean,
-      limit?: number,
-    ) => Promise<ActivityEntry[]>
-  }
-  discoveryModeRegistry?: DiscoveryModeRegistry
-  getDiscoveryConnectionSnapshot?: (userId: number) => Promise<DiscoveryConnectionSnapshot>
-  runDiscoveryMode?: (
-    request: DiscoveryModeRequest,
-    options?: { existingJobId?: number },
-  ) => Promise<{ batchId: number; artistsFound?: number }>
-  // Job recording & queries
-  jobRecorder: import('@/core/jobs/types').JobRecorder
-  jobQueries: {
-    listJobs: (
-      filters?: import('@/db/queries/jobs').ListJobsFilters,
-    ) => Promise<{ items: import('@/core/jobs/types').JobRunRow[]; total: number }>
-    getJobById: (id: number) => Promise<import('@/core/jobs/types').JobRunRow | null>
-    getJobHealth: (nextRun: Date | null) => Promise<import('@/db/queries/jobs').HealthSummary>
-    getJobsForSubscription: (
-      subId: number,
-      limit?: number,
-    ) => Promise<import('@/core/jobs/types').JobRunRow[]>
-  }
-  // Playlist deps (optional - omit in test environments without a DB)
-  playlistDeps?: PlaylistDeps
-  // Search deps (optional - absent when no search sources are configured)
-  search?: SearchDeps
-}
+// AppDependencies is the intersection of every per-domain slice in deps.ts.
+// Route files that only need a subset (see TargetDeps / LibraryDeps / etc.
+// already defined locally in some routes) can import from `./deps` instead
+// of accepting the full bag. Keeping this re-export stable avoids breaking
+// every caller.
+export type { AppDependencies } from './deps'
+
+import type { AppDependencies } from './deps'
 
 export function createApp(deps: AppDependencies) {
   const app = new Hono<HonoEnv>()
